Update 02-think-data.md
Okay, some small changes so the narrative works plus a tweak to the timings in the header. This does mean that the whole of this lesson is over three hours. But that is okay: we'll playtest it at the TNA in November and see how long it takes in reality!